+++ Controllers/Mobile/v1/AuthController.cs
@@ -1,5 +1,4 @@
 ﻿using AonFreelancing.Contexts;
-using AonFreelancing.Models;
 using AonFreelancing.Models.DTOs;
 using AonFreelancing.Models.Entities;
 using AonFreelancing.Models.Requests;
@@ -9,7 +8,6 @@ using AonFreelancing.Utilities;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
-using Client = AonFreelancing.Models.Entities.Client;
 
 namespace AonFreelancing.Controllers.Mobile.v1
 {
