--- Interfaces/IUserOperations.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace AonFreelancing.Interfaces
-{
-    public interface IUserAuthOperations
-    {
-       
-        public void Login();
-        public void Logout();
-    }
-}
