--- Interfaces/IUserPrintOperations.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace AonFreelancing.Interfaces
-{
-    public interface IUserPrintOperations
-    {
-        public void DisplayProfile();
-    }
-}
