+++ Models/DTOs/UserDTO.cs
@@ -1,5 +1,4 @@
 ﻿using System.ComponentModel.DataAnnotations;
-using System.Text.Json.Serialization;
 
 namespace AonFreelancing.Models.DTOs
 {
