+++ Models/Entities/SystemUser.cs
@@ -1,9 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations.Schema;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.ComponentModel.DataAnnotations.Schema;
+
 
 namespace AonFreelancing.Models.Entities
 {
@@ -11,10 +7,5 @@ namespace AonFreelancing.Models.Entities
     public class SystemUser : User
     {
         public string Permissions { get; set; }
-
-        //public override void DisplayProfile()
-        //{
-        //    Console.WriteLine("System User PRofile");
-        //}
     }
 }
