+++ Models/Entities/User.cs
@@ -1,19 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
-using System.Text.Json.Serialization;
-using System.Threading.Tasks;
-using AonFreelancing.Interfaces;
-using Microsoft.AspNetCore.Identity;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.AspNetCore.Identity;
 
 namespace AonFreelancing.Models.Entities
 {
-    //Replaced with 'builder.Entity<User>().HasIndex(u => u.PhoneNumber).IsUnique();' in OnModelCreate
-    //[Index(nameof(PhoneNumber), IsUnique = true)]
-
     public class User : IdentityUser<long>
     {
         public string Name { get; set; }
