+++ Models/Requests/ForgotPasswordReq.cs
@@ -4,6 +4,6 @@ namespace AonFreelancing.Models.Requests
 {
     public record ForgotPasswordReq(
         [Required, Phone, StringLength(14, MinimumLength = 14)] 
-        string? PhoneNumber
+        string PhoneNumber
     );
 }
