+++ Models/Requests/ResetPasswordReq.cs
@@ -4,12 +4,12 @@ namespace AonFreelancing.Models.Requests
 {
     public record ResetPasswordReq(
         [Required, MinLength(6, ErrorMessage = "Too short password")]
-        string? Password,
+        string Password,
     
         [Required, MinLength(6, ErrorMessage = "Too short password")]
-        string? ConfirmPassword,
+        string ConfirmPassword,
     
         [Required, Phone, StringLength(14, MinimumLength = 14)]
-        string? PhoneNumber
+        string PhoneNumber
     );
 }
